app.py
from langchain_core.prompts import ChatPromtTemplate
from langchain_core.output_parsers import StrOutputParser
from langchain_community.llms import Ollama
import streamlit as st
import os
from dotenv import load_dotenv 

load_dotenv()